logic to handle zips
